wrapper: Update Mac build script for XCode 4.1 and OS 10.7 (supports Intel Macs only)
svn path=/trunk/boinc/; revision=24231

+++ samples/wrapper/BuildMacWrapper.sh
@@ -21,7 +21,7 @@
 # Script to build Macintosh wrapper using Makefile
 #
 # by Charlie Fenton 2/15/10
-# Updated 10/11/10 for XCode 3.2 and OS 10.6 
+# Updated 11/16/11 for XCode 4.1 and OS 10.7 
 #
 ## First, build the BOINC libraries using boinc/mac_build/BuildMacBOINC.sh
 ##
@@ -33,25 +33,29 @@
 
 rm -fR ppc i386 x86_64
 
-echo
-echo "***************************************************"
-echo "********** Building PowerPC Application ***********"
-echo "***************************************************"
-echo
+if [ -d /Developer/SDKs/MacOSX10.3.sdk/ ]; then
 
-export MACOSX_DEPLOYMENT_TARGET=10.3
-export CC=/usr/bin/gcc-4.0;export CXX=/usr/bin/g++-4.0
-export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.4u.sdk,-arch,ppc"
-export VARIANTFLAGS="-arch ppc -DMAC_OS_X_VERSION_MAX_ALLOWED=1030 -DMAC_OS_X_VERSION_MIN_REQUIRED=1030 -isysroot /Developer/SDKs/MacOSX10.4u.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
+    echo
+    echo "***************************************************"
+    echo "********** Building PowerPC Application ***********"
+    echo "***************************************************"
+    echo
 
-rm -f wrapper.o
-rm -f wrapper
-make -f Makefile_mac all
+    export MACOSX_DEPLOYMENT_TARGET=10.3
+    export CC=/usr/bin/gcc-4.0;export CXX=/usr/bin/g++-4.0
+    export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.4u.sdk,-arch,ppc"
+    export VARIANTFLAGS="-arch ppc -DMAC_OS_X_VERSION_MAX_ALLOWED=1030 -DMAC_OS_X_VERSION_MIN_REQUIRED=1030 -isysroot /Developer/SDKs/MacOSX10.4u.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
 
-if [  $? -ne 0 ]; then exit 1; fi
+    rm -f wrapper.o
+    rm -f wrapper
+    make -f Makefile_mac all
 
-mkdir ppc
-mv wrapper ppc/
+    if [  $? -ne 0 ]; then exit 1; fi
+
+    mkdir ppc
+    mv wrapper ppc/
+
+fi
 
 echo
 echo "***************************************************"
@@ -59,10 +63,10 @@ echo "******* Building 32-bit Intel Application *********"
 echo "***************************************************"
 echo
 
-export MACOSX_DEPLOYMENT_TARGET=10.4
-export CC=/usr/bin/gcc-4.0;export CXX=/usr/bin/g++-4.0
-export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.4u.sdk,-arch,i386"
-export VARIANTFLAGS="-arch i386 -DMAC_OS_X_VERSION_MAX_ALLOWED=1040 -DMAC_OS_X_VERSION_MIN_REQUIRED=1040 -isysroot /Developer/SDKs/MacOSX10.4u.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
+export MACOSX_DEPLOYMENT_TARGET=10.6
+export CC=/usr/bin/gcc-4.2;export CXX=/usr/bin/g++-4.2
+export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.6.sdk,-arch,i386"
+export VARIANTFLAGS="-arch i386 -DMAC_OS_X_VERSION_MAX_ALLOWED=1040 -DMAC_OS_X_VERSION_MIN_REQUIRED=1040 -isysroot /Developer/SDKs/MacOSX10.6.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
 
 rm -f wrapper.o
 rm -f wrapper
@@ -79,11 +83,11 @@ echo "******* Building 64-bit Intel Application *********"
 echo "***************************************************"
 echo
 
-export MACOSX_DEPLOYMENT_TARGET=10.5
-export CC=/usr/bin/gcc-4.0;export CXX=/usr/bin/g++-4.0
+export MACOSX_DEPLOYMENT_TARGET=10.6
+export CC=/usr/bin/gcc-4.2;export CXX=/usr/bin/g++-4.2
 export LDFLAGS="-Wl,-arch x86_64"
-export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.5.sdk,-arch,x86_64"
-export VARIANTFLAGS="-arch x86_64 -DMAC_OS_X_VERSION_MAX_ALLOWED=1050 -DMAC_OS_X_VERSION_MIN_REQUIRED=1050 -isysroot /Developer/SDKs/MacOSX10.5.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
+export LDFLAGS="-Wl,-syslibroot,/Developer/SDKs/MacOSX10.6.sdk,-arch,x86_64"
+export VARIANTFLAGS="-arch x86_64 -DMAC_OS_X_VERSION_MAX_ALLOWED=1050 -DMAC_OS_X_VERSION_MIN_REQUIRED=1050 -isysroot /Developer/SDKs/MacOSX10.6.sdk -fvisibility=hidden -fvisibility-inlines-hidden"
 
 rm -f wrapper.o
 rm -f wrapper
